cleaning up the controllers page
